- mavenized the project
