+++ binsrc/rdf4j/src/main/java/com/openlink/virtuoso/rdf4j/driver/Version.java
@@ -20,16 +20,16 @@
  *  51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
  *
  */
-package virtuoso.rdf4j.driver;
+package com.openlink.virtuoso.rdf4j.driver;
 
-import org.eclipse.rdf4j.query.*;
+public class Version {
+    protected static final String DRV_VER = "2.0.1";
 
-public class VirtuosoTupleQuery extends VirtuosoQuery  implements TupleQuery {
-    public TupleQueryResult evaluate() throws QueryEvaluationException {
-        return null;
-    }
+    // The major and minor version number
+    protected static final String build_id = "1.0";
 
-    public void evaluate(TupleQueryResultHandler tupleQueryResultHandler) throws QueryEvaluationException, TupleQueryResultHandlerException {
-
+    public static void main(String args[])
+    {
+        System.out.println("OpenLink Virtuoso(TM) Provider for RDF4J(TM) Version " + DRV_VER + " [Build " + build_id + "]");
     }
 }
