+++ binsrc/rdf4j/src/main/java/com/openlink/virtuoso/rdf4j/driver/VirtuosoQuery.java
@@ -20,12 +20,10 @@
  *  51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
  *
  */
-package virtuoso.rdf4j.driver;
+package com.openlink.virtuoso.rdf4j.driver;
 
-import org.eclipse.rdf4j.model.Value;
-import org.eclipse.rdf4j.query.BindingSet;
-import org.eclipse.rdf4j.query.Dataset;
 import org.eclipse.rdf4j.query.Query;
+import org.eclipse.rdf4j.query.QueryInterruptedException;
 
 public class VirtuosoQuery extends VirtuosoOperation implements Query {
     /**
