+++ binsrc/rdf4j/src/main/java/com/openlink/virtuoso/rdf4j/driver/VirtuosoUpdate.java
@@ -20,13 +20,18 @@
  *  51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA
  *
  */
-package virtuoso.rdf4j.driver;
+package com.openlink.virtuoso.rdf4j.driver;
 
-import org.eclipse.rdf4j.query.BooleanQuery;
-import org.eclipse.rdf4j.query.QueryEvaluationException;
+import org.eclipse.rdf4j.query.Update;
+import org.eclipse.rdf4j.query.UpdateExecutionException;
 
-public class VirtuosoBooleanQuery extends VirtuosoQuery  implements BooleanQuery {
-    public boolean evaluate() throws QueryEvaluationException {
-        return false;
+public class VirtuosoUpdate extends VirtuosoOperation implements Update {
+
+    /**
+     * Execute this update on the repository
+     */
+    public void execute() throws UpdateExecutionException {
+
     }
+
 }
