+++ binsrc/rdf4j/src/main/resources/data.nt
